third_party/blink/web_tests/http/tests/preload/preload-media-disabled.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<link rel=preload href="../resources/test.oga" as=audio>
<link rel=preload href="../resources/test.mp4" as=video>

<script>
var t = async_test("Ensure preloads don't respect as 'video' or 'audio' until https://github.com/w3c/preload/issues/97 is resolved and implemented.");
window.addEventListener("load", t.step_func(function() {
    if (window.internals) {
        assert_false(internals.isPreloaded('../resources/test.oga'), "audio should not be preloaded.");
        assert_false(internals.isPreloaded('../resources/test.mp4'), "videos should not be preloaded.");
        t.done();
    }
}));
</script>
